requirements.txt
discord.py>=2.3.0,<3.0.0
aiosqlite>=0.20.0
